requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
psycopg2-binary
kafka-python==2.0.2
python-keycloak==3.3.0 